+++ libraries/test_session_current/src/main/java/androidx/media3/session/MediaControllerCompatProviderService.java
@@ -15,6 +15,7 @@
  */
 package androidx.media3.session;
 
+import static androidx.media3.session.RemoteMediaControllerCompat.QUEUE_IS_NULL;
 import static androidx.media3.test.session.common.CommonConstants.ACTION_MEDIA_CONTROLLER_COMPAT;
 import static androidx.media3.test.session.common.CommonConstants.KEY_ARGUMENTS;
 import static androidx.media3.test.session.common.TestUtils.SERVICE_CONNECTION_TIMEOUT_MS;
@@ -37,6 +38,7 @@ import androidx.media3.common.util.Log;
 import androidx.media3.test.session.common.IRemoteMediaControllerCompat;
 import androidx.media3.test.session.common.TestHandler;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Executor;
@@ -125,6 +127,13 @@ public class MediaControllerCompatProviderService extends Service {
       MediaControllerCompat controller = mediaControllerCompatMap.get(controllerId);
       MediaDescriptionCompat desc = (MediaDescriptionCompat) getParcelable(descriptionBundle);
       controller.removeQueueItem(desc);
+    }
+
+    @Override
+    public int getQueueSize(String controllerId) throws RemoteException {
+      MediaControllerCompat controller = mediaControllerCompatMap.get(controllerId);
+      List<MediaSessionCompat.QueueItem> queue = controller.getQueue();
+      return queue != null ? controller.getQueue().size() : QUEUE_IS_NULL;
     }
 
     @Override