+++ libraries/test_session_current/src/main/java/androidx/media3/session/RemoteMediaControllerCompat.java
@@ -49,6 +49,8 @@ import java.util.concurrent.CountDownLatch;
  * <p>Users can run {@link MediaControllerCompat} methods remotely with this object.
  */
 public class RemoteMediaControllerCompat {
+
+  public static final int QUEUE_IS_NULL = -1;
   static final String TAG = "RMediaControllerCompat";
 
   final String controllerId;
@@ -104,6 +106,10 @@ public class RemoteMediaControllerCompat {
 
   public void removeQueueItem(MediaDescriptionCompat description) throws RemoteException {
     binder.removeQueueItem(controllerId, createBundleWithParcelable(description));
+  }
+
+  public int getQueueSize() throws RemoteException {
+    return binder.getQueueSize(controllerId);
   }
 
   public void setVolumeTo(int value, int flags) throws RemoteException {